fix: Update SkillEntity and SkillItemEntity to enforce non-null skill items and adjust data types for proficiency level and rank

+++ sources/JobMagnet.Application/Contracts/Responses/Base/SkillItemBase.cs
@@ -3,8 +3,8 @@
 public sealed record SkillItemBase
 {
     public long Id { get; init; }
-    public int ProficiencyLevel { get; init; }
-    public int Rank { get; init; }
+    public ushort ProficiencyLevel { get; init; }
+    public ushort Rank { get; init; }
     public string? Name { get; init; }
     public string? Category { get; init; }
     public string? IconUrl { get; init; }
